Add files via upload
I created a parallel version of a serial algorithm that we were provided with by  M.M.Kuttel, and I used Java for the
shared memory parallel programming model . In the Abelian Sandpile, each cell contains a certain number of "grains of sand". When the number
of grains in a cell exceeds 4, in the next time step the cell "topples", distributing grains evenly to its
neighboring cells (left, right, up and down) and keeping any remainder. the program outputs to the screen the number of rows and
columns, the number of simulation timesteps required to reach the stable state and the total time
taken. I used the Fork Join Framework to parallelize the serial version
In the Abelian Sandpile, each cell contains a certain number of "grains of sand". When the number
of grains in a cell exceeds 4, in the next time step the cell "topples", distributing grains evenly to its
neighboring cells (left, right, up and down) and keeping any remainder. A picture is generated through this.